Module crypto_layer::common::crypto::algorithms::encryption

Enums

AsymmetricEncryption

Represents the available encryption algorithms.

BlockCiphers

Represents the available block cipher algorithms.

EccCurves

Specifies the curve types for Elliptic Curve Digital Signature Algorithm (ECDSA).

EccSchemeAlgorithm

Enum representing the ECC scheme interface type.

Rc2KeyBits

Specifies the key sizes for the RC2 block cipher.

StreamCiphers

Represents the available stream cipher algorithms.

SymmetricMode

Specifies the modes of operation for symmetric block ciphers.

TripleDesNumKeys

Specifies the number of keys used in Triple DES configurations.