Enum crypto_layer::common::crypto::pkcs::standards::OidType

#[repr(C)]
pub enum OidType {

    Sha1WithRsaEncryption = 0,
    Sha256WithRsaEncryption = 1,
    Sha384WithRsaEncryption = 2,
    Sha512WithRsaEncryption = 3,
    IdSha1 = 4,
    IdSha256 = 5,
    IdSha384 = 6,
    IdSha512 = 7,
    RsaEncryption = 8,
    IdRsaSsaPss = 9,
    EcPublicKey = 10,
    EcdsaWithSha1 = 11,
    EcdsaWithSha256 = 12,
    EcdsaWithSha384 = 13,
    EcdsaWithSha512 = 14,
    Aes128Cbc = 15,
    Aes192Cbc = 16,
    Aes256Cbc = 17,
    Aes128Gcm = 18,
    Aes192Gcm = 19,
    Aes256Gcm = 20,
    DhPublicNumber = 21,
    EcdhStandardCurves = 22,
    EcdhSpecifiedCurves = 23,
    SubjectKeyIdentifier = 24,
    KeyUsage = 25,
    SubjectAltName = 26,
    BasicConstraints = 27,
    ExtendedKeyUsage = 28,
    ServerAuth = 29,
    ClientAuth = 30,
    CodeSigning = 31,
    EmailProtection = 32,
    TimeStamping = 33,
    OcspSigning = 34,
    DocumentSigning = 35,
    Pkix = 36,
    Pkcs7Data = 37,
    Pkcs7SignedData = 38,
    Pkcs7EnvelopedData = 39,
    Pkcs7SignedAndEnvelopedData = 40,
    Pkcs7DigestData = 41,
    Pkcs7EncryptedData = 42,
    Pkcs9EmailAddress = 43,
    Pkcs9UnstructuredName = 44,
    Pkcs9ContentType = 45,
    Pkcs9MessageDigest = 46,
    Pkcs9SigningTime = 47,

}

Enumerates specific Object Identifier (OID) types across cryptographic functionalities.

Lists OIDs for hashing algorithms, public key cryptography, encryption algorithms, key agreement protocols, certificate extensions, extended key usage identifiers, and other notable OIDs.

Hashing Algorithms

Sha1WithRsaEncryption, Sha256WithRsaEncryption, etc.

Public Key Cryptography

RsaEncryption, EcPublicKey, EcdsaWithSha*

Encryption Algorithms

Aes128Cbc, Aes256Gcm, etc.

Key Agreement

DhPublicNumber, EcdhStandardCurves

Certificate Extensions

SubjectKeyIdentifier, ExtendedKeyUsage, etc.

Extended Key Usage OIDs

ServerAuth, ClientAuth, CodeSigning, etc.

Other notable OIDs

Pkix, Pkcs7Data, Pkcs9EmailAddress, etc.

Examples

Selecting an OID for AES 256 CBC encryption:

use tpm_poc::common::crypto::pkcs::standards::OidType;

let oid_type = OidType::Aes256Cbc;

Note

Uses #[repr(C)] for C language compatibility. Understanding and using the correct OIDs is essential for ensuring cryptographic operations adhere to the intended standards and protocols.

Variants

Sha1WithRsaEncryption = 0

Sha256WithRsaEncryption = 1

Sha384WithRsaEncryption = 2

Sha512WithRsaEncryption = 3

IdSha1 = 4

IdSha256 = 5

IdSha384 = 6

IdSha512 = 7

RsaEncryption = 8

IdRsaSsaPss = 9

EcPublicKey = 10

EcdsaWithSha1 = 11

EcdsaWithSha256 = 12

EcdsaWithSha384 = 13

EcdsaWithSha512 = 14

Aes128Cbc = 15

Aes192Cbc = 16

Aes256Cbc = 17

Aes128Gcm = 18

Aes192Gcm = 19

Aes256Gcm = 20

DhPublicNumber = 21

EcdhStandardCurves = 22

EcdhSpecifiedCurves = 23

SubjectKeyIdentifier = 24

KeyUsage = 25

SubjectAltName = 26

BasicConstraints = 27

ExtendedKeyUsage = 28

ServerAuth = 29

ClientAuth = 30

CodeSigning = 31

EmailProtection = 32

TimeStamping = 33

OcspSigning = 34

DocumentSigning = 35

Pkix = 36

Pkcs7Data = 37

Pkcs7SignedData = 38

Pkcs7EnvelopedData = 39

Pkcs7SignedAndEnvelopedData = 40

Pkcs7DigestData = 41

Pkcs7EncryptedData = 42

Pkcs9EmailAddress = 43

Pkcs9UnstructuredName = 44

Pkcs9ContentType = 45

Pkcs9MessageDigest = 46

Pkcs9SigningTime = 47

Implementations

impl OidType

pub fn as_str(&self) -> &'static str