Trait crypto_layer::common::traits::key_handle::KeyHandle

pub trait KeyHandle: Send + Sync + Debug {
    // Provided methods
    fn sign_data(&self, _data: &[u8]) -> Result<Vec<u8>, SecurityModuleError> { ... }
    fn decrypt_data(
        &self,
        _encrypted_data: &[u8]
    ) -> Result<Vec<u8>, SecurityModuleError> { ... }
    fn encrypt_data(&self, _data: &[u8]) -> Result<Vec<u8>, SecurityModuleError> { ... }
    fn verify_signature(
        &self,
        _data: &[u8],
        _signature: &[u8]
    ) -> Result<bool, SecurityModuleError> { ... }
}

Defines a common interface for cryptographic key operations.

This trait specifies methods for key operations such as signing data, encrypting, decrypting, and verifying signatures. It’s designed to be implemented by security modules that manage cryptographic keys, ensuring a consistent interface for key operations across different types of security modules. Implementors of this trait must ensure thread safety.

Provided Methods

fn sign_data(&self, _data: &[u8]) -> Result<Vec<u8>, SecurityModuleError>

Signs the given data using the cryptographic key.

Arguments

data - A byte slice representing the data to be signed.

Returns

A Result containing the signature as a Vec<u8> on success, or a SecurityModuleError on failure.

fn decrypt_data( &self, _encrypted_data: &[u8] ) -> Result<Vec<u8>, SecurityModuleError>

Decrypts the given encrypted data using the cryptographic key.

Arguments

encrypted_data - A byte slice representing the data to be decrypted.

Returns

A Result containing the decrypted data as a Vec<u8> on success, or a SecurityModuleError on failure.

fn encrypt_data(&self, _data: &[u8]) -> Result<Vec<u8>, SecurityModuleError>

Encrypts the given data using the cryptographic key.

Arguments

data - A byte slice representing the data to be encrypted.

Returns

A Result containing the encrypted data as a Vec<u8> on success, or a SecurityModuleError on failure.

fn verify_signature( &self, _data: &[u8], _signature: &[u8] ) -> Result<bool, SecurityModuleError>

Verifies the signature of the given data using the cryptographic key.

Arguments

data - A byte slice representing the data whose signature is to be verified.

signature - A byte slice representing the signature to be verified against the data.

Returns

A Result containing a boolean indicating whether the signature is valid (true) or not (false), or a SecurityModuleError on failure.